@@ -1,5 +1,5 @@
 name: hotelmanagement
-description: hotel management is an application for the management of a b & b or a horel
+description: hotel management is an application for the management of a b and b or a hotel
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
@@ -17,36 +17,36 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 environment:
-  sdk: '>=3.0.3 <4.0.0'
+  sdk: '>=3.1.5 <4.0.0'
 
 
 
 dependencies:
-  cloud_firestore: ^4.8.0
-  cupertino_icons: ^1.0.5
-  date_time_picker: ^2.1.0
-  firebase_app_check: ^0.1.4+2
-  firebase_auth: ^4.6.2
-  firebase_core: ^2.13.1
-  firebase_performance: ^0.9.2+2
+  intl: ^0.18.1
+  cloud_firestore: ^4.12.0
+  cupertino_icons: ^1.0.6
+  firebase_app_check: ^0.2.1+2
+  firebase_auth: ^4.12.0
+  firebase_core: ^2.20.0
+  firebase_performance: ^0.9.3+2
 
   flutter:
     sdk: flutter
-  flutter_facebook_auth: ^6.0.0
-  flutter_markdown: ^0.6.15
-  flutter_native_splash: ^2.3.1
-  flutter_riverpod: ^2.3.6
+  flutter_facebook_auth: ^6.0.2
+  flutter_markdown: ^0.6.18
+  flutter_native_splash: ^2.3.4
+  flutter_riverpod: ^2.4.4
   flutter_slidable: ^3.0.0
-  flutter_svg: ^2.0.6
+  flutter_svg: ^2.0.8
   flutter_switch: ^0.3.2
   g_recaptcha_v3: ^0.0.5
-  google_fonts: ^5.0.0
-  google_sign_in: ^6.1.4
-  page_transition: ^2.0.9
-  syncfusion_flutter_datepicker: ^21.2.3
+  google_fonts: ^6.1.0
+  google_sign_in: ^6.1.5
+  page_transition: ^2.1.0
+  syncfusion_flutter_datepicker: ^23.1.42
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
   flutter_test:
     sdk: flutter
 
