@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hotelmanagement/screen/loginScreen/welcome_screen.dart';
 
 class App extends StatelessWidget {
-  const App({Key? key}) : super(key: key);
+  const App({super.key});
 
   @override
   Widget build(BuildContext context) {
